DP-26714: Split GenTeaser component into a set of independent components.
Maintain backward compatibility, so that we don't have to release another major release.
